Access Unit
Created an AccessUnit class to contain all the TS packets associated with a frame including metadata, PAT and PMT.  Presently, this data type is past between decoder and the sender.  In the future, this data unit will be past to a rate limiter class.

In the main, the application will display the elapsed time of the stream file.