Improve Rate Control (#2)
* Probe Input File

Added a class to probe the input file.  For now only the duration is capture.

* Display more metadata

Query the prober for the following file properties:
- average bitrate
- metadata carriage
- metadata frequency

* Probe H.264 Stream

Create H.264 prober to query properties such as  frame/second, resolution, etc.

* Access Unit

Created an AccessUnit class to contain all the TS packets associated with a frame including metadata, PAT and PMT.  Presently, this data type is past between decoder and the sender.  In the future, this data unit will be past to a rate limiter class.

In the main, the application will display the elapsed time of the stream file.

* Rate Limiter

Created a Rate Limiter thread sits between the decoder and sender.  The class does nothing now but will be responsible for releasing packets on the network.

* Sliding Window

Implemented a rate limiter using sliding window.  The RateLimiter takes the
AccessUnit PTS and wall clock.  If the difference is with a window, output the AccessUnit.

* Change Window interval

- Change window interval base on framerate
- When building in Visual Studio, use the libraries inside the solution.
- RateLimiter implements BaseIOInterface.
- Frame count will be provided by the limiter.

* Revert changes

Revert changes in RateLimiter.

* Probe and Run

Client application needs to call Mp2tsStreamer::probe() member function before calling
Mp2tsStreamer::run().
Add new member functions to Mp2tsStreamer class so client applications can retrieve the result of a probe.
Removed unnecessary #include from files.

* Remove Rate

Removed rate argument because it is no longer needed.

* Adapt to Interface changes

Performance counter change to query the playback position during streaming.
Use the current AccessUnit PTS time that is about to be transmitted for this value.

* Starvation Issue

Sometimes the source MPEG-2 TS stream has missing PES packets which causes the sender queue to be empty.  This result in the sender thread exiting prematurely.  Stop the sender thread in the after the limiter thread joins.

* Improvements

- When reading the source file I/O errors may occur so clear the stream and try again.
- During the probe, get the file size.  The file reader use it to stop its thread.
- In the RateLimiter, improve when an AccessUnit is within the release window.

* Resolve Code Analysis Warnings

Modified code to resolve code analysis warnings.  No change in behavior.

* Update Readme

Updated readme file to stated another dependency.