Restructure task (#5)
* Interim commit

Restructure Mp2tStrm library.  Interim commit to save changes.

* Now Builds

Interim commit. The library project now builds.

* Building the App

Interim commit that builds the app.

* Building Lib

Removed unwanted files.  Rename library folder.

* Build and Test

Build the app and run the test.
Disable some debugging output to console.
Commented out Performance Counter build for now.

* Update Readme.md

Minor edits with a subsection on how to build the application.

* Performance Counter

A MSVC project uses this project for implementation.  The Mp2tStrm library depends on this MSVC project for a Performance Counter library.
The compiler macro PERFCNTR must be passed in the cmake command line to enable this dependency with the Performance Counter library.

* Updated Readme.md

Added a statement to clone the loki library.
Show basic cmake commands to build and install the app.

* Update README.md

Added Usage and Example subsections.

* Install Application

Modified CMakeList.txt file to install the application.

* Linux Build Fixes

Modified code so it can build on linux.

* Add Includes

Added includes for basic types instead of using typedefs.

* Only install the App

Commented out the commands to install the Mp2tStrm library.  Unsure it will be used by other clients.

* Add new parameters

Add -p parameter to probe the input stream and exit
Add -f parameter to set the frames per second.
The -f parameter is needed because a if the input stream is piped in, the input stream can only be read once.
So no probing is done before streaming.

* Fix Linux Build

Added missing include files to build on Linux.

* Update READMD.md

Added a To Test subsection.