Start Pause Feature (#16)
* Validate --startPosition Value

Ensure the --startPosition value is greater than zero and less the video duration.

* Add pause and start member functions

Added pause and start member fuctions to BaseIOInterface class that all derived classes must implement.
Add state information to the Mp2tStreamer class.

* Implement FSM

Implement FSM in the Mp2tStreamer class.

* Implement Pause and Start

Implemented pause and start member functions in the RateLimiter class.  Needs testing.

* Thread for reading stdin

Create a thread to read stdin.  Does nothing for now.

* Implement Pause and Start

Users can enter p on the console to pause streaming and s to restart it.
More work needs to be done because when paused the application goes into busy wait, which consumes CPU unnecessarily.

* New Test Case

Add a test case to test the --startPosition option.
Add a new sample ts file for testing.

* Runtime Commands

Added a Runtime Commands section to the Readme file.

* Avoid Busy Wait

When the application is in pause, add a second of sleep.  We wanted to avoid complicated solution using mutex and critical section.
Update Readme document.