20 fail to exit (#21)
* Non-blocking getchar

On Linux, the getchar() blocks the input thread.  Detaching the input thread fails to allow the main thread to exit.  Created a non-blocking getchar() function so the input thread can exit.

* FetchContent

Use FetchContent to download loki-lib instead of ExternalProject.

* Windows Fix

Ifdef the sleep function in input thread.

* Restore Keyboard

Restore the keyboard before the application exits.